lrtable

Struct StateGraph

pub struct StateGraph<StorageT: Eq + Hash> { /* private fields */ }

Implementations

impl<StorageT: 'static + Hash + PrimInt + Unsigned> StateGraph<StorageT>
where usize: AsPrimitive<StorageT>,

pub fn start_state(&self) -> StIdx<StorageT>

Return this state graph’s start state.

pub fn iter_stidxs(&self) -> Box<dyn Iterator<Item = StIdx<StorageT>>>

Return an iterator which produces (in order from StorageT::zero()..self.all_states_len()) all this grammar’s valid StIdxs.

pub fn closed_state(&self, stidx: StIdx<StorageT>) -> &Itemset<StorageT>

Return the itemset for closed state stidx. Panics if stidx doesn’t exist.

pub fn iter_closed_states<'a>( &'a self, ) -> Box<dyn Iterator<Item = &'a Itemset<StorageT>> + 'a>

Return an iterator over all closed states in this StateGraph.

pub fn core_state(&self, stidx: StIdx<StorageT>) -> &Itemset<StorageT>

Return the itemset for core state stidx or None if it doesn’t exist.

pub fn iter_core_states<'a>( &'a self, ) -> Box<dyn Iterator<Item = &'a Itemset<StorageT>> + 'a>

Return an iterator over all core states in this StateGraph.

pub fn all_states_len(&self) -> StIdx<StorageT>

How many states does this StateGraph contain? NB: By definition the StateGraph contains the same number of core and closed states.

pub fn edge( &self, stidx: StIdx<StorageT>, sym: Symbol<StorageT>, ) -> Option<StIdx<StorageT>>

Return the state pointed to by sym from stidx or None otherwise.

pub fn edges( &self, stidx: StIdx<StorageT>, ) -> &HashMap<Symbol<StorageT>, StIdx<StorageT>>

Return the edges for state stidx. Panics if stidx doesn’t exist.

pub fn all_edges_len(&self) -> usize

How many edges does this StateGraph contain?

pub fn pp(&self, grm: &YaccGrammar<StorageT>, core_states: bool) -> String

Pretty print this stategraph as a String. If core_states is set to true, only the core states are pretty printed; if set to false, all states (including non-core states) are pretty printed.

pub fn pp_core_states(&self, grm: &YaccGrammar<StorageT>) -> String

Return a pretty printed version of the core states, and all edges.

pub fn pp_closed_states(&self, grm: &YaccGrammar<StorageT>) -> String

Return a pretty printed version of the closed states, and all edges.